pyproject.toml
[project]
name = "knowlg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphiti-core>=0.24.0",
    "python-dotenv>=1.2.1",
]
